+++ src/reporter/cucumber/index.ts
@@ -14,6 +14,7 @@ import BaseReporter, { InternalOptions } from './base';
 import MessageReporter from './message';
 import HtmlReporter from './html';
 import JunitReporter from './junit';
+import JunitReporterModern from './junit-modern';
 import JsonReporter from './json';
 import CustomReporter, { CustomReporterOptions } from './custom';
 import { getConfigDirFromEnv } from '../../config/env';
@@ -22,6 +23,7 @@ const builtinReporters = {
   html: HtmlReporter,
   message: MessageReporter,
   junit: JunitReporter,
+  'junit-modern': JunitReporterModern, // New Junit reporter using @cucumber/junit-xml-formatter
   json: JsonReporter,
 } as const;